backend/requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
mysql-connector-python==8.0.33
PyMySQL==1.1.0
python-dotenv==1.0.0
